multi: replace IsOwner flag with OwnedScriptChecker
Remove the IsOwner boolean from VTXOIntent and VTXORequest in favor of
a data-driven OwnedScriptChecker interface. Instead of tagging each VTXO
with an ownership flag at construction time, the round FSM now queries
the owned receive scripts store at confirmation time to determine which
VTXOs to persist locally. This decouples VTXO construction from
ownership tracking and is required for OOR receive scripts to work
correctly with round VTXOs.

Key changes:
- Add OwnedScriptChecker and OwnedScriptRegistrar interfaces to round
  package
- Wire OwnedScriptChecker into ClientEnvironment and all FSM
  construction sites
- Register owned scripts in buildVTXOIntent and handleRegisterIntent
  via OwnedScriptRegistrar
- Create ownedScriptCheckerAdapter and ownedScriptRegistrarAdapter in
  darepod backed by OORArtifactPersistenceStore
- Use context.WithoutCancel in IsOwnedScript for shutdown safety
- Update all tests to use mockOwnedScriptChecker instead of IsOwner

+++ darepod/rpc_server.go
@@ -673,7 +673,7 @@ func (r *RPCServer) SendVTXO(ctx context.Context,
 				codes.InvalidArgument,
 				"recipient %d: amount must be "+
 					"between 1 and %d",
-				i, btcutil.MaxSatoshi,
+				i, int64(btcutil.MaxSatoshi),
 			)
 		}
